working on powerups...

+++ src/GameObject/PowerUp.cpp
@@ -1,11 +1,11 @@
 #include "Main.h"
 
-PowerUp::PowerUp(int id , SDL_Point position, int type) : _GameObject(position , {64,64} , 0 , STATIC )
+PowerUp::PowerUp(int id , SDL_Point position, int type) : _GameObject( position , {64,64} , 0 , STATIC )
 {
     this->id = id;
     this->type = type;
 
-    sprite = new Sprite(Game::textureManager->powerups,{160*type,0,160,160},float(0.4*Game::configuration->getScale()));
+    sprite = new Sprite(Game::textureManager->powerups,{160*type,0,160,160},0.4*Game::configuration->getScale());
 }
 
 void PowerUp::draw(int x0 , int y0)
